# packages/control-sdk/openapi/control-plane.v3.json
{
  "openapi": "3.1.0",
  "info": { "title": "Agent HQ Control Plane API", "version": "3.0.0" },
  "paths": {
...
    "/v1/executions/validate": {
      "post": {
        "operationId": "execution.validate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "caller": {
                    "properties": {
                      "servicePrincipalId": {
                        "maxLength": 64,
                        "minLength": 5,
                        "pattern": "^svc_[a-z][a-z0-9-]*$",
                        "type": "string"
                      }
                    },
                    "required": ["servicePrincipalId"],
                    "type": "object"
                  },
                  "commandId": { "pattern": "^cmd_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "contractVersion": {
                    "additionalProperties": false,
                    "properties": {
                      "major": {
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991,
                        "type": "integer"
                      },
                      "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                    },
                    "required": ["major", "minor"],
                    "type": "object"
                  },
                  "correlation": {
                    "properties": {
                      "causationCommandId": {
                        "pattern": "^cmd_[0-9A-HJKMNP-TV-Z]{26}$",
                        "type": "string"
                      },
                      "parentEventId": {
                        "pattern": "^evt_[0-9A-HJKMNP-TV-Z]{26}$",
                        "type": "string"
                      },
                      "traceId": { "pattern": "^trc_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                    },
                    "required": ["traceId"],
                    "type": "object"
                  },
                  "idempotencyKey": {
                    "maxLength": 128,
                    "minLength": 16,
                    "pattern": "^[A-Za-z0-9._:-]+$",
                    "type": "string"
                  },
                  "issuedAt": {
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
                    "type": "string"
                  },
                  "operation": { "const": "execution.validate", "type": "string" },
                  "payload": {
                    "anyOf": [
                      {
                        "properties": {
                          "agentId": {
                            "pattern": "^agt_[0-9A-HJKMNP-TV-Z]{26}$",
                            "type": "string"
                          },
                          "contextInputs": { "not": {} },
                          "contextPackage": {
                            "properties": {
                              "compilerVersion": {
                                "maxLength": 64,
                                "minLength": 1,
                                "type": "string"
                              },
                              "contentDigest": {
                                "pattern": "^sha256:[a-f0-9]{64}$",
                                "type": "string"
                              },
                              "contextPackageId": {
                                "pattern": "^ctx_[0-9A-HJKMNP-TV-Z]{26}$",
                                "type": "string"
                              },
                              "schemaVersion": {
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991,
                                "type": "integer"
                              }
                            },
                            "required": [
                              "contextPackageId",
                              "contentDigest",
                              "schemaVersion",
                              "compilerVersion"
                            ],
                            "type": "object"
                          },
                          "outputContractRef": {
                            "maxLength": 512,
                            "minLength": 1,
                            "type": "string"
                          },
                          "policySnapshot": {
                            "properties": {
                              "contentDigest": {
                                "pattern": "^sha256:[a-f0-9]{64}$",
                                "type": "string"
                              },
                              "policySnapshotId": {
                                "maxLength": 256,
                                "minLength": 1,
                                "type": "string"
                              },
                              "revision": {
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991,
                                "type": "integer"
                              }
                            },
                            "required": ["policySnapshotId", "revision", "contentDigest"],
                            "type": "object"
                          },
                          "profileVersionId": {
                            "pattern": "^pfv_[0-9A-HJKMNP-TV-Z]{26}$",
                            "type": "string"
                          },
                          "projectState": {
                            "properties": {
                              "projectId": {
                                "pattern": "^prj_[0-9A-HJKMNP-TV-Z]{26}$",
                                "type": "string"
                              },
                              "revision": {
                                "maximum": 9007199254740991,
                                "minimum": 0,
                                "type": "integer"
                              },
                              "workspaceId": {
                                "pattern": "^wsp_[0-9A-HJKMNP-TV-Z]{26}$",
                                "type": "string"
                              }
                            },
                            "required": ["workspaceId", "projectId", "revision"],
                            "type": "object"
                          },
                          "runtimeRequirements": {
                            "items": {
                              "maxLength": 128,
                              "minLength": 1,
                              "pattern": "^[a-z][a-z0-9.-]*$",
                              "type": "string"
                            },
                            "maxItems": 128,
                            "type": "array"
                          },
                          "skillVersionIds": {
                            "items": {
                              "pattern": "^skv_[0-9A-HJKMNP-TV-Z]{26}$",
                              "type": "string"
                            },
                            "maxItems": 128,
                            "type": "array"
                          },
                          "taskId": { "pattern": "^tsk_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                        },
                        "required": [
                          "taskId",
                          "agentId",
                          "profileVersionId",
                          "skillVersionIds",
                          "projectState",
                          "policySnapshot",
                          "runtimeRequirements",
                          "outputContractRef",
                          "contextPackage"
                        ],
                        "type": "object"
                      },
                      {
                        "properties": {
                          "agentId": {
                            "pattern": "^agt_[0-9A-HJKMNP-TV-Z]{26}$",
                            "type": "string"
                          },
                          "contextInputs": {
                            "additionalProperties": false,
                            "properties": {
                              "budgets": {
                                "additionalProperties": false,
                                "properties": {
                                  "maximumBytes": {
                                    "exclusiveMinimum": 0,
                                    "maximum": 9007199254740991,
                                    "type": "integer"
                                  },
                                  "maximumTokens": {
                                    "exclusiveMinimum": 0,
                                    "maximum": 9007199254740991,
                                    "type": "integer"
                                  }
                                },
                                "required": ["maximumBytes", "maximumTokens"],
                                "type": "object"
                              },
                              "candidates": {
                                "items": {
                                  "additionalProperties": false,
                                  "properties": {
                                    "itemId": {
                                      "pattern": "^psi_[0-9A-HJKMNP-TV-Z]{26}$",
                                      "type": "string"
                                    },
                                    "itemRevision": {
                                      "exclusiveMinimum": 0,
                                      "maximum": 9007199254740991,
                                      "type": "integer"
                                    },
                                    "priority": {
                                      "maximum": 9007199254740991,
                                      "minimum": -9007199254740991,
                                      "type": "integer"
                                    },
                                    "required": { "type": "boolean" }
                                  },
                                  "required": ["itemId", "itemRevision", "required", "priority"],
                                  "type": "object"
                                },
                                "maxItems": 10000,
                                "type": "array"
                              },
                              "objective": { "maxLength": 16384, "minLength": 1, "type": "string" },
                              "returnContract": {
                                "additionalProperties": false,
                                "properties": {
                                  "contractRef": {
                                    "maxLength": 512,
                                    "minLength": 1,
                                    "type": "string"
                                  }
                                },
                                "required": ["contractRef"],
                                "type": "object"
                              },
                              "successCriteria": {
                                "items": { "maxLength": 4096, "minLength": 1, "type": "string" },
                                "maxItems": 128,
                                "minItems": 1,
                                "type": "array"
                              }
                            },
                            "required": [
                              "objective",
                              "candidates",
                              "successCriteria",
                              "returnContract",
                              "budgets"
                            ],
                            "type": "object"
                          },
                          "contextPackage": { "not": {} },
                          "outputContractRef": {
                            "maxLength": 512,
                            "minLength": 1,
                            "type": "string"
                          },
                          "policySnapshot": {
                            "properties": {
                              "contentDigest": {
                                "pattern": "^sha256:[a-f0-9]{64}$",
                                "type": "string"
                              },
                              "policySnapshotId": {
                                "maxLength": 256,
                                "minLength": 1,
                                "type": "string"
                              },
                              "revision": {
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991,
                                "type": "integer"
                              }
                            },
                            "required": ["policySnapshotId", "revision", "contentDigest"],
                            "type": "object"
                          },
                          "profileVersionId": {
                            "pattern": "^pfv_[0-9A-HJKMNP-TV-Z]{26}$",
                            "type": "string"
                          },
                          "projectState": {
                            "properties": {
                              "projectId": {
                                "pattern": "^prj_[0-9A-HJKMNP-TV-Z]{26}$",
                                "type": "string"
                              },
                              "revision": {
                                "maximum": 9007199254740991,
                                "minimum": 0,
                                "type": "integer"
                              },
                              "workspaceId": {
                                "pattern": "^wsp_[0-9A-HJKMNP-TV-Z]{26}$",
                                "type": "string"
                              }
                            },
                            "required": ["workspaceId", "projectId", "revision"],
                            "type": "object"
                          },
                          "runtimeRequirements": {
                            "items": {
                              "maxLength": 128,
                              "minLength": 1,
                              "pattern": "^[a-z][a-z0-9.-]*$",
                              "type": "string"
                            },
                            "maxItems": 128,
                            "type": "array"
                          },
                          "skillVersionIds": {
                            "items": {
                              "pattern": "^skv_[0-9A-HJKMNP-TV-Z]{26}$",
                              "type": "string"
                            },
                            "maxItems": 128,
                            "type": "array"
                          },
                          "taskId": { "pattern": "^tsk_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                        },
                        "required": [
                          "taskId",
                          "agentId",
                          "profileVersionId",
                          "skillVersionIds",
                          "projectState",
                          "policySnapshot",
                          "runtimeRequirements",
                          "outputContractRef",
                          "contextInputs"
                        ],
                        "type": "object"
                      }
                    ]
                  },
                  "payloadHash": { "pattern": "^[a-f0-9]{64}$", "type": "string" },
                  "projectId": { "pattern": "^prj_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "requestId": { "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" },
                  "workspaceId": { "pattern": "^wsp_[0-9A-HJKMNP-TV-Z]{26}$", "type": "string" }
                },
                "required": [
                  "caller",
                  "contractVersion",
                  "requestId",
                  "workspaceId",
                  "correlation",
                  "commandId",
                  "idempotencyKey",
                  "payloadHash",
                  "operation",
                  "issuedAt",
                  "payload"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "contractVersion": {
                      "additionalProperties": false,
                      "properties": {
                        "major": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                      },
                      "required": ["major", "minor"],
                      "type": "object"
                    },
                    "correlation": {
                      "additionalProperties": false,
                      "properties": {
                        "causationCommandId": {},
                        "parentEventId": {},
                        "traceId": {}
                      },
                      "required": ["traceId"],
                      "type": "object"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "executionPlan": {
                          "additionalProperties": false,
                          "properties": {
                            "contentDigest": {
                              "pattern": "^sha256:[a-f0-9]{64}$",
                              "type": "string"
                            },
                            "executionPlanId": {}
                          },
                          "required": ["executionPlanId", "contentDigest"],
                          "type": "object"
                        },
                        "valid": { "const": true, "type": "boolean" }
                      },
                      "required": ["valid", "executionPlan"],
                      "type": "object"
                    },
                    "requestId": {}
                  },
                  "required": ["contractVersion", "requestId", "correlation", "data"],
                  "type": "object"
                }
              }
            },
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "contractVersion": {
                      "additionalProperties": false,
                      "properties": {
                        "major": {
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991,
                          "type": "integer"
                        },
                        "minor": { "maximum": 9007199254740991, "minimum": 0, "type": "integer" }
                      },
                      "required": ["major", "minor"],
                      "type": "object"
                    },
                    "correlation": {
                      "additionalProperties": false,
                      "properties": {
                        "causationCommandId": {},
                        "parentEventId": {},
                        "traceId": {}
                      },
                      "required": ["traceId"],
                      "type": "object"
                    },
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "class": {
                          "enum": [
                            "validation",
                            "authentication",
                            "authorization",
                            "conflict",
                            "stale_reference",
                            "capability_mismatch",
                            "runtime_unavailable",
                            "internal"
                          ],
                          "type": "string"
                        },
                        "code": {
                          "maxLength": 96,
                          "minLength": 1,
                          "pattern": "^[A-Z][A-Z0-9_]*$",
                          "type": "string"
                        },
                        "details": {
                          "additionalProperties": {},
                          "propertyNames": { "maxLength": 64, "type": "string" },
                          "type": "object"
                        },
                        "message": { "maxLength": 512, "minLength": 1, "type": "string" },
                        "remediation": { "maxLength": 256, "minLength": 1, "type": "string" },
                        "retryable": { "type": "boolean" },
                        "source": {
                          "enum": [
                            "client",
                            "auth",
                            "policy",
                            "runtime",
                            "provider",
                            "persistence",
                            "workflow",
                            "system"
                          ],
                          "type": "string"
                        }
                      },
                      "required": ["class", "code", "message", "retryable", "source"],
                      "type": "object"
                    },
                    "requestId": {}
                  },
                  "required": ["contractVersion", "requestId", "correlation", "error"],
                  "type": "object"
                }
              }
            },
            "description": "Normalized error response"
          }
        },
        "security": [{ "serviceBearer": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "service-credential" }
    }
  }
}